Convert ElementEditor to ViewComponent
Replace the ElementEditor decorator and _element.html.erb partial with
a single Alchemy::Admin::ElementEditor ViewComponent.

This consolidates view-specific logic and template rendering into one
component, following the pattern established by ingredient editor
components.
