+++ app/views/alchemy/admin/elements/index.html.erb
@@ -38,11 +38,11 @@
           id="main-content-elements"
           style="--padding: 0"
         >
-          <%= render @elements.map { |element| Alchemy::ElementEditor.new(element) } %>
+          <%= render Alchemy::Admin::ElementEditor.with_collection(@elements) %>
         </sl-tab-panel>
         <% @fixed_elements.each do |element| %>
           <sl-tab-panel id="fixed_element_<%= element.id %>" name="fixed-element-<%= element.id %>" style="--padding: 0" class="scrollable-elements">
-            <%= render Alchemy::ElementEditor.new(element) %>
+            <%= render Alchemy::Admin::ElementEditor.new(element: element) %>
           </sl-tab-panel>
         <% end %>
       </sl-tab-group>
@@ -53,7 +53,7 @@
         data-droppable-elements="<%= @page.element_definition_names.join(' ') %>"
         data-element-name="main-content-elements"
       >
-        <%= render @elements.map { |element| Alchemy::ElementEditor.new(element) } %>
+        <%= render Alchemy::Admin::ElementEditor.with_collection(@elements) %>
       </div>
     <% end %>
   </alchemy-elements-window>
